added features news component to news page
